Refactor request options and some related types
Moves common request options e.g., `ContentLength` into `typespec_client_core` while reducing the number of seldom-used (specified directly) types exported from `typespec_client_core::http`.
