+++ sdk/typespec/typespec_client_core/src/http/request/mod.rs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+pub mod options;
+
 #[cfg(not(target_arch = "wasm32"))]
 use crate::stream::SeekableStream;
 use crate::{
