+++ sdk/typespec/typespec_client_core/src/http/request/options/content_length.rs
@@ -1,8 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-use crate::headers::{self, Header};
+use crate::http::headers::{self, Header};
+#[cfg(doc)]
+use crate::http::request::Request;
 
+/// Set the "content-length" header in a [`Request`].
 #[derive(Debug, Clone, Copy)]
 pub struct ContentLength(i32);
