+++ sdk/typespec/typespec_client_core/src/http/request/options/if_match_condition.rs
@@ -1,9 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-use crate::headers::{self, Header};
+use crate::http::headers::{self, Header};
+#[cfg(doc)]
+use crate::http::request::Request;
 use headers::{IF_MATCH, IF_NONE_MATCH};
 
+/// Sets the "if-match" or "if-none-match" headers in a [`Request`].
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum IfMatchCondition {
     Match(String),
