+++ sdk/typespec/typespec_client_core/src/http/request/options/if_modified_since_condition.rs
@@ -1,12 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+#[cfg(doc)]
+use crate::http::request::Request;
 use crate::{
-    date,
-    headers::{self, Header, HeaderName},
+    date::{self, OffsetDateTime},
+    http::headers::{self, Header, HeaderName},
 };
-use typespec_client_core::date::OffsetDateTime;
 
+/// Sets the "if-modified-since" or "if-unmodified-since" headers in a [`Request`].
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum IfModifiedSinceCondition {
     Modified(OffsetDateTime),
